AwesomeFruits/AwesomeFruits.WebAPI.Users/Controllers/AuthController.cs
using System;
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;
using System.Text;
using System.Threading.Tasks;
using AwesomeFruits.Application.DTOs;
using AwesomeFruits.Application.Interfaces;
using AwesomeFruits.Domain.Exceptions;
using AwesomeFruits.WebAPI.Users.Constants;
using AwesomeFruits.WebAPI.Users.Exceptions;
using AwesomeFruits.WebAPI.Users.Model;
using AwesomeFruits.WebAPI.Users.Validations;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Configuration;
using Microsoft.IdentityModel.Tokens;

namespace AwesomeFruits.WebAPI.Users.Controllers;

[Route("api/[controller]")]
[ApiController]
public class AuthController : ControllerBase
{
    private readonly IConfiguration _configuration;
    private readonly IUserService _userService;

    public AuthController(IUserService userService, IConfiguration configuration)
    {
        _userService = userService;
        _configuration = configuration;
    }

    [HttpPost("register")]
    public async Task<ActionResult<SaveUserDto>> Register(SaveUserDto saveUserDto)
    {
        try
        {
            ValidateSaveUserDtoRequest(saveUserDto);

            return Created("", await _userService.SaveUserAsync(saveUserDto));
        }
        catch (Exception ex)
        {
            switch (ex)
            {
                case UserNameAlreadyExistsException:
                    return BadRequest(ex.Message);
                case ValidationErrorsException:
                    throw;
                default:
                    return StatusCode(500, ResponseConstants.InternalError);
            }
        }
    }

    [HttpPost("login")]
    public async Task<ActionResult<Token>> Login(LoginUserDto login)
    {
        try
        {
            var user = await _userService.GetValidLoggedInUser(login);

            var token = GenerateJwtToken(user);
            var accessToken = new Token { AccessToken = token };

            return Ok(accessToken);
        }
        catch (Exception ex)
        {
            if (ex is UserCredentialsNotValidException) return Unauthorized();

            return StatusCode(500, ResponseConstants.InternalError);
        }
    }

    private string GenerateJwtToken(UserDto user)
    {
        var securityKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["Jwt:Key"]));
        var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256);

        var claims = new[]
        {
            new Claim(JwtRegisteredClaimNames.Sub, user.Id.ToString()),
            new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString())
        };

        var token = new JwtSecurityToken(
            _configuration["Jwt:Issuer"],
            _configuration["Jwt:Audience"],
            claims,
            expires: DateTime.Now.AddHours(1),
            signingCredentials: credentials);

        return new JwtSecurityTokenHandler().WriteToken(token);
    }

    private static void ValidateSaveUserDtoRequest(SaveUserDto request)
    {
        var errors = new SaveUserDtoValidator().ValidateSaveUserDto(request);

        if (errors.Count > 0) throw new ValidationErrorsException(errors);
    }
}